oorpb: move oorwire package to rpc/oorpb
Move the OOR mailbox wire protocol package from the top-level oorwire/
directory into rpc/oorpb/ to colocate it with the other RPC proto
packages (roundpb, etc.) under the rpc/ tree. This rename also changes
the Go package name from oorwire to oorpb for consistency with the
naming convention used by roundpb.

All proto stubs are regenerated via make rpc with the updated
go_package option pointing to rpc/oorpb, and all import sites are
updated accordingly.
