+++ rpc/oorpb/oorwire.proto
@@ -1,8 +1,8 @@
 syntax = "proto3";
 
-package oorwire;
+package oorpb;
 
-option go_package = "github.com/lightninglabs/darepo-client/oorwire";
+option go_package = "github.com/lightninglabs/darepo-client/rpc/oorpb";
 
 // OORMailboxService defines unary OOR mailbox methods used in the toy E2E
 // connector path.
@@ -12,7 +12,8 @@ service OORMailboxService {
     rpc SubmitPackage (SubmitPackageRequest) returns (SubmitPackageResponse);
 
     // FinalizePackage submits finalized checkpoint PSBTs for a session.
-    rpc FinalizePackage (FinalizePackageRequest) returns (FinalizePackageResponse);
+    rpc FinalizePackage (FinalizePackageRequest)
+        returns (FinalizePackageResponse);
 }
 
 // OOROutPoint is a protobuf representation of wire.OutPoint.
