+++ rpc/oorpb/payloads_test.go
@@ -1,4 +1,4 @@
-package oorwire
+package oorpb
 
 import (
 	"bytes"
